Make serialization default to @attributes
In instances where we're serializing an object with an @attributes
instance variable defined, we want to default to using that value. This
is present when using ActiveModel::Attributes.

Some use cases of ActiveModel::Serialization just define an attributes
method returning a hash, so we want to use that method when @attributes
isn't available.