Add fallback host for SystemTestCase driven by RackTest
For links to work in SystemTestCases, it requires the `host` to be added
to the `default_url_options`. Since
78c734386cddf5ee533ad1915d3769b9cc1be3a2 this is done by looking at
`Capybara.app_host` or `Capybara.current_session.server_url`.

This works correctly when using the Selenium webdriver, as Capybara sets
either the `Capybara.app_host` or the `Capybara.current_session.server_url`.
However the Capybara RackTest driver uses paths instead of URL's, so it doesn't
set Capybara.app_host or Capybara.current_session.server_url.

For SystemTestCase to work for the RackTest driver a host has to be
available for url options. We can fallback to "http://127.0.0.1" which
was the default before 78c734386cddf5ee533ad1915d3769b9cc1be3a2.