Avoid instance_exec for controller callbacks
When a proc is passed to a callback (either as a condition or as the
callback itself) it is evaluated using instance_exec on the controller
instance. Calling instance_exec creates a new singleton class for the
object, which then will require new inline method caches.

This commit avoids creating the extra controller singleton classes when
:only or :except are passed to a callback.