Fix migration compatibility to create SQLite references/belongs_to column as a integer when migration version is 6.0
Since #38717 the `ReferenceDefinition` wasn't being
instantiated with the column type as `:integer` in
the compatibility module when the adapter is SQLite
for 6.0 migrations which make such columns be created
as `bigint` instead of `integer`. Differing from
the way they are created in prior versions or newer
versions, which is always `integer`.

Fixes #43168