Deprecate `[]` for non-scalar parameters (nested parameters and Arrays)
Calling `Parameters#[]` can have unexpected results as it allows access
to the unfiltered parameters.
For example a parameter might be used in a view:

    params = ActionController::Parameters.new(query: "ruby!")
    "You searched for: #{params[:query]}"
    # => "You searched for: ruby!"

When the parameter is a nested parameter, the inspect of the parameter
is outputted instead.

    params = ActionController::Parameters.new(query: { name: "ruby!" })
    "You searched for: #{params[:query]}"
    # => "You searched for: <ActionController::Parameters {"name"=>"ruby!"} permitted: false>

By not allowing `[]` for non-scalar types, we can prevent the problem.
For backwards compatibility to get the non-scalar parameter, the `param_at` method is added.

Calling `[]` on scalar types still works.

    params = ActionController::Parameters.new(query: "ruby!")
    params[:query] => # "ruby!"

Calling `[]` on nested params will no longer be permitted:

    params = ActionController::Parameters.new(
      person: {
        name: { "Matz" }
      }
    )
    # Directly calling params[:person] isn't permitted, as it isn't a
    # scalar type.
    params[:person] # => nil

    # Permitting the param still works
    params.require(:person).permit(:name)[:name] # => "Matz"

    # The old behaviour still works with `param_at`
    params.param_at(:person)[:name] # => "Matz"

Fixes #42942