Add granular deduplication for Callbacks.
Prior to this commit, callback deduplication used a cachkey composed of filter and kind for all parents and children.  Where kind is on of (before, after or around) and filter is the method called on the object.  This filtering assumed that the load order of defining callbacks would always be Parent then Children.  This is a good assumption because the vast majority of cases will follow that pattern.

However, this assumption causes an interesting state when we end up in this scenario
 - include a module on a parent that defines callbacks at runtime after all class initiation completes
 - that new callback has the same filter and kind as a child classes callback

When this happen the childs callback will be detected as a duplicate and deleted, then replaced with the parents callback.  This is an unusual scenario but it would cause unexpected issues for proc and if/unless constraints that are defined in the child because the parent will override them.  Imagine a scenario where the parent says only: :index but the child says only: :show.  The parent would override the childes conditions to run the callback.  The load order issue is similar with skip_callback.

This only impacts callbacks defined with a Symbol.  Callbacks which define a Class that expose methods are not impacted.

Deduplication was added to prevent duplicate callbacks in the same class and when defining a symbol.

Based on the above, this change aligns Class based and symbol callbacks.  It prevents duplicate callbacks in parents from overriding children by introducing a target instance variable that is used to scope the deduplication to the class that is defining it.  Finally, we handle cases for skip_callback whenever we end up in a situation with mulitple callbacks.