Keep connection alive on TransactionRollbackError
Prior to this change, databases that raised a TransactionRollbackError
(either through ActiveRecord::Deadlock or ActiveRecord::SerializationFailure)
would have their connections removed from the pool and discarded.

This commit prevents the connection from being thrown away if the
transaction itself was marked as invalidated. This is consistent with
similar logic that keeps the connection in the pool if the transaction
successfully rolled back.

Fixes #43130