Expose whether the schema cache was loaded from a dump
We're not currently using this internally, but by exposing it, we allow
an interested application to confirm the expected schema dump was used
(the file was found, the schema version matched, etc).

For applications with particularly large schemas, especially when
combined with large process pools, failure to locate and load a dump
needs to be handled as a startup error; the usual fallback of starting
with an empty cache can cause further issues.