Add ability to define connections in a yaml file
For some applications, mainly ones using a tenant sharding strategy, may
want to be able to define the connections outside of the model. A good
example of this is when your database is sharded by customer and each
customer gets their own "tenant" but the application is deployed to a
server where 100 customers are using. In those cases we can't define the
connections in a model because it will be different for each server.

This PR aims to solve that problem by providing a mechanism for reading
a file (`config/connections.yml`) that defines the connections for each
"tenant" as a shard. The file can be checked into the application or can
be generated on deploy providing a way to have a more dynamic set of
connections for an application.

To use this functionality applications can create a
`config/connections.yml` that defines shards that correspond to each
tenant.

Applications first create a database.yml:

```yaml

production:
  primary:
    database: tenant_router_db
  primary_replica:
    database: tenant_router_db
    replica: true
  database_one:
    database: my_db_one
  database_one_replica:
    database: my_db_one
    replica: true
  database_two:
    database: my_db_two
  database_two_replica:
    database: my_db_two
    replica: true
```

Then they create a connections.yml:

```yaml
ApplicationRecord:
  shards:
    default:
      writing: primary
      reading: primary
    customer_one:
      writing: database_one
      reading: database_one_replica
    customer_two:
      writing: database_two
      reading: database_two_replica
```

Lastly, in `ApplicationRecord` we add the following:

```
class ApplicationRecord < ActiveRecord::Base
  primary_abstract_class

  connects_to_file
end
```

The call to `connects_to_file` allows for an elegant way for different deploy
targets to have unique connections defined without polluting the models. This
can also be used for applications that have many shards and they don't
want to define `connects_to` in the model.

The functionality is pretty simple. If a `connections.yml` file exists
Rails will parse the file and create `ConnectionConfig` objects. When
`connects_to_file` is called then Active Record finds the
`ConnectionConfig` for the class and passes the required arguments to
`connects_to`.

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>