Implement `form_for` by delegating to `form_with`
Improve the parity between `form_for` and `form_with` by implementing
`form_for` in terms of `form_with`.

Replaces `html_options` transformations with coercion of data into the
shape it needs to be in order to delegate to `form_with`.

In the same spirit, this commit also implements `fields_for` in terms of
`fields`.