[styles.css] Add theme styles and generator
Adds color styles for light and dark themes, and a generator method for
combining them.

The `color.css.erb` in particular is generated because multiple sections
of it are reused, so this allows it to stay DRY while still providing
the eventual use case of being able to add a theme switcher.