Enable eager loading by default on CI systems
This is a pattern we've been using for ages at Shopify.

When you are running test locally, most of the time you
run only a subset, so it's better to load as little code
as possible to have a faster time to first test result.

But when you are on CI, it's usally much preferable to
eager load the whole application because you will likely
need all the code anyway, and even if the test suite is
split across runners, it's preferable to load
all the code to ensure any codefile that may have side
effects is loaded.

This also ensure that if some autoloaded constants are
not properly tested on CI, at least they'll be loaded
and obvious errors (e.g. SyntaxError) will be caught on
CI rather than during deploy.