Proof of concept: ActiveSupport::ExecutionContext
Many places in Active Support and Rails in general use `Thread.current#[]`
to store "request (or job) local data". This often cause problems with
`Enumerator` because it runs in a different fiber.

Ref: https://github.com/rails/rails/pull/38905
Ref: https://github.com/rails/rails/pull/39428
Ref: https://github.com/rails/rails/pull/34495

But the required locality depends on how the application is deployed.
If you use `falcon`, then you want fiber-locals, if you use `puma/sidekiq`
you want thread-locals, and if you use `unicorn/resque`, you can do with
thread-locals, but could also just as well do with process globals.

Since we keep having to deal this thread vs fiber local problem in various
places, I think it suggest we have the need for a proper API to store
some state in the current unit of work scope.

Then it would b up to the application owner to set the proper isolation
level based on how they deploy their Rails app.

For backward compatibility it could ship with `:fiber` isolation as a default
but longer term `:thread` would make more sense as it would work fine for
all deployment targets except `falcon`.