Don't `SELECT * FROM information_schema.tables`
Depending on MySQL version and configuration, this is more than two
orders of magnitude slower than selecting only the two columns we need
for this query, which don't require reading any of the data files or
recalculating innodb stats.

https://dev.mysql.com/doc/refman/5.6/en/information-schema-optimization.html

Performance regression was introduced in #39712, before then the query
took around 40-50ms on my test environment, after it took 4-8s. With
this change it's back to 40-50ms.