Extract annotations using a parser for Ruby files
When you're looking for annotations, it's possible (albeit extremely unlikely) for it to pick up things that aren't comments (see the added test for an example). If we instead use a parser like ripper, we can filter down to just the comments and only try to extract the annotations from the source file's comments.