Fix a NoMethodError in associate_records_from_unscoped
Prior to 2d988d03fcd5a3952e4fefeed9a8f760e71c37cb, the method
`associate_records_from_unscoped` would silently ignore nil values
that were passed in as records. With that change, it raises a new
exception when the `[]` method is called on each of the records.
This commit restores the behavior to silently drop nils.