Fix a regression in association preloader
2d988d0 changed Associations::Preloader::Batch to no longer group
records by class in order to work with records of differing STI
classes. This introduced a problem where the preloader may attempt
to load an association on a record that doesn't have it, if two
records of differing classes shared the same `id`.

This commit restores grouping to fix the regression, but groups by
base class rather than class, so that it will work with records using
STI like 2d988d0 intended.